Note added...
Not happy.

There seems to be no way whatsoever to STOP Maxim from re-fetching an
audio that has already been fetched to play a note that is still
playing, let alone one that has stopped playing a while before.

P.js preloading only works for images, and base64 data has no guarantee
to work for audio, and will be HUGE, so not feasible.

And the notes themselves still have gaps, which I have to edit.